+++ lib/multiselect_dropdown.dart
@@ -2,7 +2,6 @@ library multiselect_dropdown;
 
 export 'src/class/chip_config.dart';
 export 'src/class/controller.dart';
-export 'src/class/value_item.dart';
 export 'src/enum/app_enums.dart';
 export 'src/multiselect_dropdown.dart';
 export 'src/widgets/hint_text.dart';